// src/app/models/Region.ts
export class Region {
    // id: 3967
    // creationDate: "2018-08-20T10:33:07.9041484"
    // modificationDate: "2018-08-20T10:33:07.9041484"
    // isDeleted: false
    id?: number;
    creationDate?: Date;
    modificationDate?: Date;
    isDeleted?: boolean;
    geoNameId: number;
    name: string;
    featureCode: string;
    countryId: number;
    regionCode: string;
    parentRegionId: string;

}